+++ mrb_doc/image.rb
@@ -1,7 +1,0 @@
-# Returns an Image object with the screen data.
-# @return [Image]
-def get_screen_data
-  # mrb_get_screen_data
-  # src/mruby_integration/image.cpp
-  Image.new
-end